requirements.txt
kafka-python==2.0.2
requests==2.31.0
streamlit>=1.32.0
streamlit-autorefresh>=1.0.1
pandas>=2.0.0
